fix Z3_Backend.hpp(329): warning C4244: 'initializing': conversion from 'const unsigned __int64' to 'unsigned int'

+++ src/metaSMT/BitBlast.hpp
@@ -763,10 +763,10 @@ namespace metaSMT {
         }
 
         result_type operator() (bvtags::bvsint_tag , boost::any arg ) {
-          typedef boost::tuple< int64_t, uint64_t> P;
+          typedef boost::tuple< int64_t, unsigned> P;
           P p = boost::any_cast<P>(arg);
           int64_t value = boost::get<0>(p);
-          uint64_t width = boost::get<1>(p);
+          unsigned width = boost::get<1>(p);
         
           bv_result ret (width);
           result_base one  = _solver(predtags::true_tag (), boost::any());
