+++ src/metaSMT/backend/CVC4.hpp
@@ -176,10 +176,10 @@ namespace metaSMT {
       }
 
       result_type operator()( bvtags::bvsint_tag , boost::any arg ) {
-        typedef boost::tuple<int64_t, uint64_t> Tuple;
+        typedef boost::tuple<int64_t, unsigned> Tuple;
         Tuple tuple = boost::any_cast<Tuple>(arg);
         int64_t value = boost::get<0>(tuple);
-        uint64_t width = boost::get<1>(tuple);
+        unsigned width = boost::get<1>(tuple);
 
         ::CVC4::BitVector bvValue (width, ::CVC4::Integer(value));
         return exprManager_.mkConst(bvValue);
